Add check for already running command
At the start and after each processed file we now write a value to the
appconfig. This value is cleared on a successfull exit. If another run
is started before the current run is finished we check if the latest
activity is >30 minutes ago. If so we assume the process is dead.

Signed-off-by: Roeland Jago Douma <roeland@famdouma.nl>