Fix some spelling mistakes in readme
I would habe merged this with the other pull request (https://github.com/rullzer/previewgenerator/pull/55) but didn't found a way to do so. I am new to github. Is it the usual way to offer pull requests to a repo where you are not contributor/have no write access to do the "create a fork and edit the file" method? Or is there some more elegant way?