Do not insert already queued items
Fixes #85

When a file is touched again before the cron job is ran the file is just
added to the database again. This makes no sense. As no new preview will
be generated.

It doesn't do much harm besides increasing the DB size and having the
preview generation command run a bit longer (note it does not generate
them multiple times).

In the future (NC 13+) we should switch to migrations. We can then
adjust the table to just have a unique key on both by filtering out
doubles first etc.

Signed-off-by: Roeland Jago Douma <roeland@famdouma.nl>